+++ core/_cli/seeds/03-roles.sql
@@ -1,3 +1,0 @@
-INSERT INTO public.roles(id, name, description) VALUES (1, 'Admin', '');
-INSERT INTO public.roles(id, name, description) VALUES (2, 'Member', '');
-INSERT INTO public.roles(id, name, description) VALUES (3, 'Guest', '');