+++ core/stream-source-files/stream.int.test.js
@@ -25,6 +25,7 @@ async function commonSetup () {
   const fileExtension = { id: 1, value: '.wav' }
   await models.FileExtension.create(fileExtension)
   const project = (await models.Project.findOrCreate({ where: { id: 'foo', name: 'my project', createdById: seedValues.primaryUserId } }))[0]
+  await models.UserProjectRole.create({ user_id: project.createdById, project_id: project.id, role_id: seedValues.roleOwner })
   const stream = await models.Stream.create({ id: 'j123k', name: 'Jaguar Station', latitude: 10.1, longitude: 101.1, createdById: seedValues.primaryUserId, projectId: project.id })
   return { audioFileFormat, audioCodec, fileExtension, project, stream }
 }