+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 ### Bug Fixes
 * **core**: Fix Arbimon API path to separated the path between staging as `legacy-api` and production as `api`
 * **core**: Fix logging request error on endpoint `streams/:id/detections` receiving array of objects
+* **core**: Add `Create Owner to role and role permission table` migration files
+* **core**: Add `include_roles` and `permissions` params to `project/:id/users` endpoint
+* **core**: Fix test cases after above changes
 
 ## 1.3.4 (2023-12-xx)
 ### Bug Fixes